requirements.txt
Django>=5.0
psycopg2==2.9.10
django-extensions
pillow
ipdb

httpie
djangorestframework